refactor(schema): move StandardSchema utilities from converters/ to standard/
Separate field schema converters from general StandardSchema utilities.
converters/ now exclusively handles field schema transformations (to-arktype,
to-drizzle), while standard/ contains utilities for arbitrary StandardSchema
types used in action inputs.
